Reduce RPC column request limit (#47)
* Reduce MAX_REQUEST_DATA_COLUMN_SIDECARS const to 512
* Fix propagation of call exception to returned future
* Fix/refactor SimpleSidecarRetriever to respect maxRequestCount and the number of currently running requests per peer
* Fix compile error from prev PR

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/networking/libp2p/rpc/DataColumnSidecarsByRangeRequestMessage.java
@@ -41,8 +41,7 @@ public class DataColumnSidecarsByRangeRequestMessage
           namedSchema("start_slot", SszPrimitiveSchemas.UINT64_SCHEMA),
           namedSchema("count", SszPrimitiveSchemas.UINT64_SCHEMA),
           namedSchema(
-              "columns",
-              SszUInt64ListSchema.create(specConfigEip7594.getNumberOfColumns().intValue())));
+              "columns", SszUInt64ListSchema.create(specConfigEip7594.getNumberOfColumns())));
     }
 
     @Override
